requirements.txt
snowflake-snowpark-python[pandas]
streamlit==1.26.0
openai==0.27.8
langchain
pandas
pydantic
snowflake-connector-python
snowflake-sqlalchemy
flask-sqlalchemy
supabase
unstructured
tiktoken
black
replicate
boto3
prompts
package_name
streamlit_chat
langchain_experimental
pinecone-client
python-dotenv
sqlalchemy
chromadb
dotenv